# contracts/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AccessForge API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/workspaces/{workspace_id}/journey-capabilities": {
      "get": {
        "description": "What a journey may ask for, from the domain's own allowlists.\n\nServed rather than duplicated in the client, because a client-side copy is a second definition\nof the policy and it will disagree the day the allowlist changes. An author offered a control\nthe server refuses is being invited to fail.",
        "operationId": "journey_capabilities_v1_workspaces__workspace_id__journey_capabilities_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Journey Capabilities V1 Workspaces  Workspace Id  Journey Capabilities Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journey Capabilities",
        "tags": [
          "journeys"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/journeys": {
      "post": {
        "description": "Compile and seal a new immutable journey version.\n\n201, not 202: unlike a run, this finishes inside the request. The response carries the digests\nthat were actually sealed, so a caller can show a reviewer what freezing committed to rather\nthan a re-derivation of it.",
        "operationId": "freeze_journey_version_v1_workspaces__workspace_id__journeys_post",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Freeze Journey Version V1 Workspaces  Workspace Id  Journeys Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Journey Version",
        "tags": [
          "journeys"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/journeys/{journey_version_id}": {
      "get": {
        "operationId": "get_journey_version_v1_workspaces__workspace_id__journeys__journey_version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "journey_version_id",
            "required": true,
            "schema": {
              "title": "Journey Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Journey Version V1 Workspaces  Workspace Id  Journeys  Journey Version Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Journey Version",
        "tags": [
          "projects"
        ]
      }
    },
    "/v1/workspaces/{workspace_id}/journeys/{journey_version_id}/policy": {
      "get": {
        "description": "The navigator policy for a frozen version, exactly as sealed.\n\nServed so a reviewer can read what the navigator will be given. It contains no oracle material\nby construction \u2014 the compiler builds it that way \u2014 and showing it is how that claim becomes\ncheckable by a person rather than asserted in a document.",
        "operationId": "journey_navigator_policy_v1_workspaces__workspace_id__journeys__journey_version_id__policy_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "journey_version_id",
            "required": true,
            "schema": {
              "title": "Journey Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Journey Navigator Policy V1 Workspaces  Workspace Id  Journeys  Journey Version Id  Policy Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Journey Navigator Policy",
        "tags": [
          "journeys"
        ]
      }
    },
...
    "/v1/workspaces/{workspace_id}/projects/{project_id}/journeys": {
      "get": {
        "operationId": "list_journey_versions_v1_workspaces__workspace_id__projects__project_id__journeys_get",
        "parameters": [
          {
            "in": "path",
            "name": "workspace_id",
            "required": true,
            "schema": {
              "title": "Workspace Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Journey Versions V1 Workspaces  Workspace Id  Projects  Project Id  Journeys Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Journey Versions",
        "tags": [
          "journeys"
        ]
      }
    },
...
  }
}